fix: allow the placement of pods on control-plane nodes by default
aligns tolerations with helm-chart defaults

On EKS, when using a mix of both Node Groups and Fargate Profiles,
the falcon-operator custom resource tries to install the DaemonSet
pods on all nodes, not just EC2 Node Groups. By default, the
operator should create a DaemonSet that does not deploy to Fargate
Nodes.