Set the Falcon sensor cleanup node sleep time to infinity
The node cleanup DaemonSet cleans up a Falcon sensor installation
in a pod init-container, with the pod container just doing a noop
sleep for 10 seconds.

In clusters with a large number of nodes, it's possible that the
cleanup DaemonSet runs to completion and enters CrashLoopBackOff
on a particular node, before it has been deployed to all nodes.

By modifying the sleep time to infinity, the pods remain in the
"Running" state instead of "CrashLoopBackOff".

See https://github.com/CrowdStrike/falcon-helm/pull/395 for more
information.